+++ orury-client/src/main/java/org/orury/client/user/application/UserService.java
@@ -1,4 +1,4 @@
-package org.orury.domain.user.domain;
+package org.orury.client.user.application;
 
 import org.orury.domain.user.domain.dto.UserDto;
 import org.springframework.web.multipart.MultipartFile;
